construction-ai-langgraph/requirements.txt
# requirements.txt
# LangChain核心
langchain==0.0.346
langgraph==0.0.13
langchain-community==0.0.10
langchain-openai==0.0.2

# 模型集成
openai==1.3.0
transformers==4.35.0
torch==2.1.0
sentencepiece==0.1.99
accelerate==0.24.1

# OCR相关
paddleocr==2.7.0.3
paddlepaddle==2.5.1
pillow==10.1.0
opencv-python==4.8.1.78
pdf2image==1.16.3

# Web框架
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# 数据处理
pandas==2.1.3
numpy==1.24.3
pydantic==2.5.0
pydantic-settings==2.1.0

# 工具库
redis==5.0.1
prometheus-client==0.19.0
python-json-logger==2.0.7
tqdm==4.66.1

# 开发工具
black==23.11.0
isort==5.12.0
pytest==7.4.3
pytest-asyncio==0.21.1